Finally succeeded in debugging all the hanging errors in the code for so long. Now, the code runs perfectly. Just have to check with downloaded data for the filters. Checked the code with dummy data and it is working perfectly well. (The combination of MATLAB and C skiiping the mex business with file handling and proper pointers and 2D array usage is working now). Things to follow are : Checking the code with downloaded data. Rectifying the code and cleaning it (it looks very bad right now). Also, some things such as the size has been hardcoded so that maybe needs to change.

+++ High_Precision_Noise_Estimation/filter_forms_modified.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 
-void iir_df2_double(double* input, long double* output, int s, double **sos, double g, int n)
+void iir_df2_double(double* input, double* output, int s, double **sos, double g, int n)
 {
     int i,j;
     double **w = malloc(n*sizeof(double*));
